# Conflicts:
#	backend/src/controllers/user.controller.js
#	backend/src/routes/user.route.js
#	backend/src/services/user.service.js
#	mobile/lib/api/user_api.service.dart
#	mobile/lib/main.dart
#	mobile/lib/screens/home/home_screen.dart

+++ mobile/lib/core/base/base_screen.dart
@@ -110,8 +110,11 @@ abstract class BaseScreen<T extends StatefulWidget> extends State<T> {
 
   Future afterFirstBuild(BuildContext context) async {}
 
-  void simpleBottomSheet(
-      {required Widget child, String? title, int height = 350}) {
+  void simpleBottomSheet({
+    required Widget child,
+    String? title,
+    double height = 350,
+  }) {
     showModalBottomSheet(
       context: context,
       isScrollControlled: true,
